# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LEVH API",
    "description": "Local-first memory layer for AI agents and humans",
    "version": "2.32.0"
  },
  "paths": {
...
    "/api/v1/agents/connect": {
      "post": {
        "summary": "Agent Connect",
        "description": "Record an agent connecting to LEVH.",
        "operationId": "agent_connect_api_agents_connect_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_session_id}/heartbeat": {
      "post": {
        "summary": "Agent Heartbeat",
        "description": "Send a heartbeat to keep an agent connection alive.",
        "operationId": "agent_heartbeat_api_agents__agent_session_id__heartbeat_post_v1",
        "parameters": [
          {
            "name": "agent_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentHeartbeatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_session_id}/disconnect": {
      "post": {
        "summary": "Agent Disconnect",
        "description": "Disconnect an agent from LEVH.",
        "operationId": "agent_disconnect_api_agents__agent_session_id__disconnect_post_v1",
        "parameters": [
          {
            "name": "agent_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDisconnectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "get": {
        "summary": "List Agents",
        "description": "List all agent connections (active and disconnected).",
        "operationId": "list_agents_api_agents_get_v1",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSessionOut"
                  },
                  "title": "Response List Agents Api Agents Get V1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/collaboration/{project}": {
      "get": {
        "summary": "Agent Collaboration",
        "description": "Get collaboration info for agents working on the same project.",
        "operationId": "agent_collaboration_api_agents_collaboration__project__get_v1",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaborationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkpoints": {
      "post": {
        "summary": "Create Checkpoint",
        "description": "Create a checkpoint of current work state.",
        "operationId": "create_checkpoint_api_checkpoints_post_v1",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Checkpoints",
        "description": "List recent checkpoints.",
        "operationId": "list_checkpoints_api_checkpoints_get_v1",
        "parameters": [
          {
            "name": "agent_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Agent Name"
            }
          },
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentCheckpointOut"
                  },
                  "title": "Response List Checkpoints Api Checkpoints Get V1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentCheckpointOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "checkpoint_type": {
            "type": "string",
            "title": "Checkpoint Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "memory_ids_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Ids Json"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_name",
          "checkpoint_type",
          "created_at"
        ],
        "title": "AgentCheckpointOut",
        "description": "An ``agent_checkpoints`` row (``memory_ids_json`` parsed client-side\nstays out of the way; the raw column is what the handler returns)."
      },
      "AgentConnectRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "project": {
            "type": "string",
            "title": "Project",
            "default": ""
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentConnectRequest"
      },
      "AgentConnectResponse": {
        "properties": {
          "agent_session_id": {
            "type": "string",
            "title": "Agent Session Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "connected_at": {
            "type": "string",
            "title": "Connected At"
          }
        },
        "type": "object",
        "required": [
          "agent_session_id",
          "agent_name",
          "display",
          "icon",
          "connected_at"
        ],
        "title": "AgentConnectResponse",
        "description": "``POST /api/agents/connect`` - the created agent-session record."
      },
      "AgentDisconnectResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "agent_session_id": {
            "type": "string",
            "title": "Agent Session Id"
          },
          "disconnected_at": {
            "type": "string",
            "title": "Disconnected At"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "agent_session_id",
          "disconnected_at"
        ],
        "title": "AgentDisconnectResponse"
      },
      "AgentHeartbeatResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "agent_session_id": {
            "type": "string",
            "title": "Agent Session Id"
          },
          "last_heartbeat": {
            "type": "string",
            "title": "Last Heartbeat"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "agent_session_id",
          "last_heartbeat"
        ],
        "title": "AgentHeartbeatResponse"
      },
...
      "AgentSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "agent_display": {
            "type": "string",
            "title": "Agent Display"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "connected_at": {
            "type": "string",
            "title": "Connected At"
          },
          "last_heartbeat_at": {
            "type": "string",
            "title": "Last Heartbeat At"
          },
          "disconnected_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disconnected At"
          },
          "metadata_json": {
            "type": "string",
            "title": "Metadata Json",
            "default": "{}"
          },
          "online": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_name",
          "agent_display",
          "status",
          "connected_at",
          "last_heartbeat_at"
        ],
        "title": "AgentSessionOut",
        "description": "An ``agent_sessions`` row as the API returns it: the raw row (``metadata_json`` stays the stored\nJSON string) and ``online`` computed per request from the in-process\npresence table."
      },
...
      "CheckpointCreateResponse": {
        "properties": {
          "checkpoint_id": {
            "type": "string",
            "title": "Checkpoint Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "checkpoint_id",
          "agent_name",
          "title",
          "created_at"
        ],
        "title": "CheckpointCreateResponse",
        "description": "``POST /api/checkpoints`` - the created checkpoint's identity."
      },
      "CheckpointRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name",
            "default": "unknown"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Work checkpoint"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "project": {
            "type": "string",
            "title": "Project",
            "default": ""
          },
          "checkpoint_type": {
            "type": "string",
            "title": "Checkpoint Type",
            "default": "manual"
          },
          "memory_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Memory Ids"
          }
        },
        "type": "object",
        "title": "CheckpointRequest"
      },
      "CollaborationResponse": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project"
          },
          "agents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Agents"
          },
          "shared_checkpoints": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shared Checkpoints"
          },
          "collaboration_score": {
            "type": "integer",
            "title": "Collaboration Score"
          }
        },
        "type": "object",
        "required": [
          "project",
          "collaboration_score"
        ],
        "title": "CollaborationResponse",
        "description": "``GET /api/agents/collaboration/{project}``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}